requirements.txt
langchain==0.3.26
sentence-transformers==4.1.0
pypdf==5.6.1 
python-dotenv==1.1.0
langchain-pinecone==0.2.8 
langchain-community==0.3.26
langchain-huggingface==0.1.0  
requests==2.32.3  

streamlit==1.39.0  
flask==3.1.1 
gunicorn==20.1.0

-e .  # for this setup.py file